Fixed error with A6, A7 pin plus couple major changes
The pin A6 & A7 cannot be used for anything digital on Nano. So removed
Also moved the bluetooth pins and removed buttons for sets to get some free pin.  Made the bluetooth module powered by 3.3V so we don't need to use voltage divider. Also made the 9V available so we can use it for powering all the score controllers